requirements.txt
customtkinter>=5.2.0
ollama>=0.1.0
requests>=2.31.0
psutil>=5.9.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0
